Code_Blazers-Hackzion-main/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-doctr[pytorch]
requests
python-multipart
pillow
regex
torch
torchvision
supabase
python-dotenv
